Data layer with jpa completed

+++ 351002/Alexeichik/MyStartRestApplication/src/main/java/com/github/Lexya06/startrestapp/service/abstraction/BaseEntityService.java
@@ -1,19 +1,30 @@
 package com.github.Lexya06.startrestapp.service.abstraction;
 
 import com.github.Lexya06.startrestapp.model.entity.abstraction.BaseEntity;
+import com.github.Lexya06.startrestapp.model.entity.realization.Label;
+import com.github.Lexya06.startrestapp.model.repository.impl.MyCrudRepositoryImpl;
+import com.github.Lexya06.startrestapp.service.customexception.MyEntitiesNotFoundException;
 import com.github.Lexya06.startrestapp.service.customexception.MyEntityNotFoundException;
 import com.github.Lexya06.startrestapp.service.mapper.impl.GenericMapperImpl;
-import com.github.Lexya06.startrestapp.model.repository.abstraction.MyCrudRepository;
+import com.querydsl.core.types.Predicate;
+import lombok.Getter;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.querydsl.binding.QuerydslPredicate;
 
+import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public abstract class BaseEntityService<T extends BaseEntity, RequestDTO, ResponseDTO> {
+    @Getter
     Class<T> entityClass;
     public BaseEntityService(Class<T> entityClass) {
         this.entityClass = entityClass;
     }
     // abstractions to reduce code count
-    protected abstract MyCrudRepository<T> getRepository();
+    protected abstract MyCrudRepositoryImpl<T> getRepository();
     protected abstract GenericMapperImpl<T,RequestDTO,ResponseDTO> getMapper();
     public ResponseDTO createEntity(RequestDTO requestDTO) {
         T entity = getMapper().createEntityFromRequest(requestDTO);
@@ -28,13 +39,15 @@ public abstract class BaseEntityService<T extends BaseEntity, RequestDTO, Respon
         return getMapper().createResponseFromEntity(entity);
     }
 
-    public Set<ResponseDTO> getEntities() {
-        Set<T> entities = getRepository().findAll();
+    public List<ResponseDTO> getEntities(Predicate predicate, Pageable pageable) {
+        List<T> entities = getRepository().findAll(predicate, pageable).getContent();
         return getMapper().createResponseFromEntities(entities);
     }
 
     public void deleteEntityById(Long id) {
-        getRepository().findById(id).orElseThrow(()->new MyEntityNotFoundException(id, entityClass));
+        if (!getRepository().existsById(id)) {
+            throw new MyEntityNotFoundException(id, entityClass);
+        }
         getRepository().deleteById(id);
     }
 
@@ -42,4 +55,13 @@ public abstract class BaseEntityService<T extends BaseEntity, RequestDTO, Respon
         T entity = getRepository().findById(id).orElseThrow(()->new MyEntityNotFoundException(id, entityClass));
         return getMapper().createResponseFromEntity(entity);
     }
+
+
+    public T getEntityReferenceWithCheckExistingId(Long id) {
+        if (!getRepository().existsById(id)) {
+            throw new MyEntityNotFoundException(id, entityClass);
+        }
+        return getRepository().getReferenceById(id);
+    }
+
 }
